Implemented data reader and more low-levels
Data reader written, needs testing.

Implemented all of the motor instructions (e.g. ping, reset, action, reg
write, etc.), although not everything is for N parameters (the general
case).

Sync write still needs to be written. An array & its size can be used to
pass in the variable-length arguments

Also, an error handler has been created to parse status packet error
codes. This needs to be completed.

Also, the data writer could be made more robust by including code that
requests and parses a status packet, to ensure the message was received
as intended. If it was not received properly, the program could re-send
the message