Major refactor of motor code base
Integrated AX12A and  MX28 libraries.

DynamixelV1IO contains all the common low-level interfacing for AX12A and MX28 motors. That is, this is where the instruction set is implemented, so this can be thought of as the bottom "layer".

DynamixelProtocolV1 implements the common register-setting functions for AX12A and MX28 motors. That is, this is where all the common functionality for the two motors is contained. This "layer" makes calls to DynamixeV1IO

Then, AX12A and MX28 .c/.h contain the unique macros and functions for each motor. Notably, this is compliance margin/slope setting for the AX12A, and P I D gain setting for the MX28


Former-commit-id: d47da30640110432245228046c5244ae28b3c67d [formerly 46fd513fd819d873074b6ed7680ae8b255e69f2a]
Former-commit-id: 58c446684b215beddc080d8e8947c42b79b0a6ca