Made all OS objects static, and moved a few things for comm around
Made OS objects static because it seems that the APIs work differently this way (specifically for binary semaphores). Basically, when dynamically allocated, semaphores are initialized with a count of 1, and when statically allocated, semaphores are initialized with a count of 0. For our purposes, the second case is more convenient, although of course a call to take or give in either case would make them equivalent.

Static allocation also ensures that all the resources are system would need are guaranteed as long as everything compiles, basically. This is good for our application.


Former-commit-id: 54cadf1df78b5d0f5718153d16eda3d5dcfc62c7 [formerly f86962d98d872d5a4d18e550b43e74ecbd1aba9e]
Former-commit-id: a74768084319786ffc6b4f9c8ac5b5afdcedbd98