Corrected numerous small errors and tested this with a simple MCU program
The test was not thorough by any means. Just ran this on the MCU:

  uint8_t buff2[84] = {0};
  while(1){
     HAL_UART_Receive(&huart2, buff2, sizeof(buff2), 100);
     HAL_UART_Transmit(&huart2, buff2, sizeof(buff2), 100);
  }

This code commit addressed several problems in the last version that had to do with how types were passed through functions and manipulated. Also, the try-catch block was causing problems rather than handling them ;) so a future improvement would be getting those in so soften any runtime blows