Update project includes, function names & minor formatting changes
Had to update the include paths for the project to rely on code in the Robot project. Revised a few of the #include directives in various files, and renamed the functions to begin with "OS_" so that they don't have the same call signature as the actual FreeRTOS functions. Minor reformatting as well.

It compiles fine now


Former-commit-id: 753eb06687b68028e052b42630fc1daa93dd5554 [formerly 6d89f60cd8fcd27035621b1ce88dbe9bbab96881]
Former-commit-id: 199669dadd67a1e3b4578118b37520d0bf31c6c3