Update project includes, function names & minor formatting changes
Had to update the include paths for the project to rely on code in the Robot project. Revised a few of the #include directives in various files, and renamed the functions to begin with "OS_" so that they don't have the same call signature as the actual FreeRTOS functions. Minor reformatting as well.

It compiles fine now