Fix incorrect logic in UartDriver transmit and receive & update tests accordingly
Previously, the tests did not modify the value saved in the pointer variable called notification, when OS_xTaskNotifyWait was called. Now this is taken into account in the most basic sense. In the future we could also add a test that makes sure transmit/receive will return false if non-blocking IO is used and the received notification does not have the correct value


Former-commit-id: dcc6a9e2710fb2abdef31bd5ec32c9845aabad77 [formerly 2df5ada203a97da7ab6ee0d54605a404f92930f8]
Former-commit-id: 3f2499aa71f9839be53eaa16f53602bc0662db1c