Fix incorrect logic in UartDriver transmit and receive & update tests accordingly
Previously, the tests did not modify the value saved in the pointer variable called notification, when OS_xTaskNotifyWait was called. Now this is taken into account in the most basic sense. In the future we could also add a test that makes sure transmit/receive will return false if non-blocking IO is used and the received notification does not have the correct value