SystemConf.h: make UART handle declarations macros
We should not declare data in the header file; this
caused a linkage problem on the MinGW compiler
for unit tests. Also a short term fix to get unit tests
building again.

Ideally, unit tests should never need to include
SystemConf.h. We will work removing that dependency
as a separate isssue.