SystemConf.h: make UART handle declarations macros
We should not declare data in the header file; this
caused a linkage problem on the MinGW compiler
for unit tests. Also a short term fix to get unit tests
building again.

Ideally, unit tests should never need to include
SystemConf.h. We will work removing that dependency
as a separate isssue.


Former-commit-id: 41726930db6dd54a565204f7db0ddb8279d0a7d3 [formerly 4d23b3248537b3de1655603b94e26e9ef9707fd0]
Former-commit-id: ba69eb6a827f6bc4930cad36e37150c5040f5ee8