Working UART muncher
Much simpler and actually works with moderate (~100Hz) packet rates
at the expense of worse latency in the face of drift (as we no longer
expect wraparound in the buffer)

There is still a lingering problem on Nucleo-32 boards when the
publisher is faster than it's consumed in the chip. I think the USB
handler has an internal buffer and overflow that trips and holds the
CTS upstream. Not sure what data rate this happens exactly -- this
solution may only work solely because it is faster as I have yet to
identify a problem with this updated version of the first attempt.