Allow more bytes than expected in JX UART handler
Before when too many bytes came in vs. expectation the handler would err
when setting the delay for the remainder of the packet to a negative
value. Surround with try/catch to stop the thread from dying.

This probably means we drop packets, I bet it's from a new packet coming
in and being received in the state machine. Until that becomes a problem
given the failure rate seems to be in the order of several seconds so
far we can probably take the packet loss.