Reinforce UART FSM against clobbering due to OS UART latency
Because of the one-wire UART used in the JX servo chain, all read and
write routines must be done carefully to avoid clobbering each other's
data by releasing control too early from a read block to a write.

Previously, this was done assuming the OS buffering time was short, and
we had very slim timeouts margining the expected synchronous
send/receive of packets. We now know that there are significant delays
up to a few ms by the kernel to release data from the RW buffers on the
wires and to the program.

The write delay is easy enough to account for, because the one-wire
creates an echo on the line to listen for, so once we read the echo we
know we're past the write delay.

The read delay is trickier, because the signal from the reads is usually
much worse due to the transmission line length, and we have to expect a
fairly high rate of frame drops. This means that we can't easily rely on
a signal from the read side on when we expect no more responses on the
line and can reliably release control to a writing block. The
alternative is timing. We have measured UART tick times via roundtrips
of the loopback. Both a consumer PC (AMD Ryzen) and Jetson Nano running
Ubuntu (20 and 18 resp.) show quantization of read timeouts to 1ms (see
write and read delay times in `soccer_hardware_bez3/proc/`) so the
expected max wait time due solely to the kernel delay is probably on the
order of ~10ms. This data has been used to create timings to abandon the
read safely if the last servo does not appear to have responded yet.