Add per-port lock to JX servo transact, notes about timeout tests
The serial IO is still not entirely stable. In the previous commit,
using timeout=0 and with locks outside of jx_servo_util.uart_transact
the transmissions would abruptly end after a while. I chalked this up to
the blind 128-byte reads and/or possible inconsistent state as Serial
read/write is not thread-safe. Both are odd explanations - in the first,
Serial.timeout = 0 is documented to return immediately after the poll.
Note I had timeouts in the previous commit, so this is clearly blocking
on a Serial.read. On the other hand, I've definitely protected every
uart_transact by a lock I've tacked onto the serial object. Furthermore,
the read and write politely handing off control to each other for
thousands of iterations really flies in the face of this.

So I still don't actually have an answer for the hanging. I've
definitely seen it happen, and it didn't trigger timeouts, which are
checked unconditionally in the loops, and there are no possibly infinite
loops in the uart_transact of the previous commit. Now, with extra
paranoia I've cut the byte counts and used non-zero finite timeouts
everywhere, and also surrounded the uart_transact with path-specific
locks. If the serial objects are cloned, at least they'll still have the
same path, so this condition is stronger than indexing by serial
objects. It's still sketchy since there is a global lock table in the JX
file -- raising the question if threads release locks in all cases when
they die -- and affirming why I would rather have locks packaged with
serial objects. But for now, I've run this through hours of tests and it
hasn't hung again.

One other mysterious behavior still happening is that increasing
"SILENT_READ_TIMEOUT" doesn't seem to be helping spurious packet fails
where nothing is read at all. Many of my theories for why this is
possible have fallen flat. The FTDI-232 USB-UART has a 128-byte receive
buffer, definitely big enough for all of the messages. The bytes that
are sent on the wire aren't being corrupted according to the scope
traces. I don't see a longer timeout helping the fail rate at all
(qualitatively). So the only two explanations are:

- It drops outgoing packets (very bad, I hope we can at least trust the
  serial module not to do that)
- The kernel is truly so busy that it ignores UART for at least 60ms
  (but yet is able to run the thread to notice the timeout?). This seems
  unlikely and is also bad if it's true.

I'll mention these fails are definitely dependent on the CPU load on the
master -- switching applications back and forth causes more fails of
both the active (>1 byte read) and silent (0 bytes read) timeouts.