Fix imu buffer chomping
Buffer was previously completely mismanaged, probably not even ever
being cleared. Also, the data pipeline is host-limited, i.e. the host
can't consume the data quickly enough, so asking for 28 bytes at a time
just balloons the system UART buffer and the program falls behind.
Instead, extend that to a read that exhausts the buffer so we have the
latest bytes all the time.